Circle and RoundedRectangle now use each a different logic for drawing -> each made a subclass of PulsatorLayout.
