+++ pulsator4droid/src/main/res/values/attrs.xml
@@ -1,21 +1,16 @@
 <resources>
     <declare-styleable name="Pulsator4Droid">
-        <attr name="pulse_count" format="integer" />
-        <attr name="pulse_duration" format="integer" />
-        <attr name="pulse_color" format="color" />
-        <attr name="pulse_repeat" format="integer" />
-        <attr name="pulse_maxScale" format="float" />
-        <attr name="pulse_startFromScratch" format="boolean" />
+        <attr name="pulse_count" format="integer"/>
+        <attr name="pulse_duration" format="integer"/>
+        <attr name="pulse_color" format="color"/>
+        <attr name="pulse_repeat" format="integer"/>
+        <attr name="pulse_maxScale" format="float"/>
+        <attr name="pulse_startFromScratch" format="boolean"/>
         <attr name="pulse_interpolator" format="enum">
-            <enum name="Linear" value="0" />
-            <enum name="Accelerate" value="1" />
-            <enum name="Decelerate" value="2" />
-            <enum name="AccelerateDecelerate" value="3" />
-        </attr>
-        <!-- Correlates to constants defined in PulsatorLayout (e.g. PULSE_TYPE_CIRCLE) -->
-        <attr name="pulse_shape" format="enum">
-            <enum name="circle" value="0" />
-            <enum name="rounded_rectangle" value="1" />
+            <enum name="Linear" value="0"/>
+            <enum name="Accelerate" value="1"/>
+            <enum name="Decelerate" value="2"/>
+            <enum name="AccelerateDecelerate" value="3"/>
         </attr>
     </declare-styleable>
 </resources>
